Fix nested entering in the namespace
This change ensures that if the terminal isn't wide enough
for the regular ps -ef command to spell out the full command
of the process the script doesn't break. Without the "-w"
switch ("ignore width"), the command line would become
truncated and "grep" would fail to find an existing systemd
process.

Signed-off-by: Paul-Stelian Olaru <olarupaulstelian97@gmail.com>